+++ src/main/java/kr/hhplus/be/server/coupon/infrastructure/CouponPolicyRepositoryImpl.java
@@ -1,7 +1,8 @@
 package kr.hhplus.be.server.coupon.infrastructure;
 
 import kr.hhplus.be.server.coupon.domain.CouponPolicy;
-import kr.hhplus.be.server.coupon.domain.CouponPolicyRepository;
+import kr.hhplus.be.server.coupon.domain.repository.CouponPolicyRepository;
+import kr.hhplus.be.server.coupon.infrastructure.repository.CouponPolicyJpaRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
