+++ src/main/java/kr/hhplus/be/server/order/infrastructure/repository/OrderItemJpaRepository.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.order.infrastructure;
+package kr.hhplus.be.server.order.infrastructure.repository;
 
 import kr.hhplus.be.server.order.domain.OrderItem;
 import org.springframework.data.jpa.repository.JpaRepository;
