+++ src/main/java/kr/hhplus/be/server/order/infrastructure/repository/OrderJpaRepository.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.order.infrastructure;
+package kr.hhplus.be.server.order.infrastructure.repository;
 
 import kr.hhplus.be.server.order.domain.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
