+++ src/main/java/kr/hhplus/be/server/payment/infrastructure/PaymentRepositoryImpl.java
@@ -1,7 +1,8 @@
 package kr.hhplus.be.server.payment.infrastructure;
 
 import kr.hhplus.be.server.payment.domain.Payment;
-import kr.hhplus.be.server.payment.domain.PaymentRepository;
+import kr.hhplus.be.server.payment.domain.repository.PaymentRepository;
+import kr.hhplus.be.server.payment.infrastructure.repository.PaymentJpaRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
