+++ src/main/java/kr/hhplus/be/server/product/application/ProductService.java
@@ -4,8 +4,8 @@ import kr.hhplus.be.server.common.exception.EntityNotFoundException;
 import kr.hhplus.be.server.common.exception.StockRecoveryException;
 import kr.hhplus.be.server.order.domain.OrderItem;
 import kr.hhplus.be.server.product.domain.Product;
-import kr.hhplus.be.server.product.domain.ProductRepository;
-import kr.hhplus.be.server.product.domain.TopProductRepository;
+import kr.hhplus.be.server.product.domain.repository.ProductRepository;
+import kr.hhplus.be.server.product.domain.repository.TopProductRepository;
 import kr.hhplus.be.server.product.domain.TopProductView;
 import kr.hhplus.be.server.product.domain.dto.response.ProductDetailResponse;
 import kr.hhplus.be.server.product.domain.dto.response.ProductResponse;
