--- src/main/java/kr/hhplus/be/server/product/domain/TopProductRepository.java
@@ -1,7 +1,0 @@
-package kr.hhplus.be.server.product.domain;
-
-import java.util.List;
-
-public interface TopProductRepository {
-    List<TopProductView> findAll();
-}
