+++ src/main/java/kr/hhplus/be/server/product/domain/repository/ProductRepository.java
@@ -1,7 +1,7 @@
-package kr.hhplus.be.server.product.domain;
+package kr.hhplus.be.server.product.domain.repository;
 
 
-import kr.hhplus.be.server.product.domain.dto.response.TopProductResponse;
+import kr.hhplus.be.server.product.domain.Product;
 
 import java.util.List;
 import java.util.Optional;
