+++ src/main/java/kr/hhplus/be/server/product/infrastructure/repository/ProductJpaRepository.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.product.infrastructure;
+package kr.hhplus.be.server.product.infrastructure.repository;
 
 import kr.hhplus.be.server.product.domain.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
