+++ src/main/java/kr/hhplus/be/server/product/infrastructure/repository/TopProductJpaRepository.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.product.infrastructure;
+package kr.hhplus.be.server.product.infrastructure.repository;
 
 import kr.hhplus.be.server.product.domain.TopProductView;
 import org.springframework.data.jpa.repository.JpaRepository;
