+++ src/test/java/kr/hhplus/be/server/coupon/application/CouponServiceIntegrationTest.java
@@ -3,8 +3,8 @@ package kr.hhplus.be.server.coupon.application;
 import kr.hhplus.be.server.TestcontainersConfiguration;
 import kr.hhplus.be.server.coupon.domain.Coupon;
 import kr.hhplus.be.server.coupon.domain.CouponPolicy;
-import kr.hhplus.be.server.coupon.infrastructure.CouponJpaRepository;
-import kr.hhplus.be.server.coupon.infrastructure.CouponPolicyJpaRepository;
+import kr.hhplus.be.server.coupon.infrastructure.repository.CouponJpaRepository;
+import kr.hhplus.be.server.coupon.infrastructure.repository.CouponPolicyJpaRepository;
 import kr.hhplus.be.server.testdata.CouponTestDataLoader;
 import kr.hhplus.be.server.user.domain.User;
 import kr.hhplus.be.server.user.infrastructure.UserJpaRepository;
