+++ src/test/java/kr/hhplus/be/server/coupon/application/CouponServiceTest.java
@@ -3,8 +3,8 @@ package kr.hhplus.be.server.coupon.application;
 import kr.hhplus.be.server.coupon.domain.Coupon;
 import kr.hhplus.be.server.coupon.domain.CouponPolicy;
 
-import kr.hhplus.be.server.coupon.domain.CouponPolicyRepository;
-import kr.hhplus.be.server.coupon.domain.CouponRepository;
+import kr.hhplus.be.server.coupon.domain.repository.CouponPolicyRepository;
+import kr.hhplus.be.server.coupon.domain.repository.CouponRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
