+++ src/test/java/kr/hhplus/be/server/payment/application/PaymentIntegrationTest.java
@@ -4,11 +4,9 @@ import jakarta.persistence.EntityManager;
 import kr.hhplus.be.server.TestcontainersConfiguration;
 import kr.hhplus.be.server.payment.domain.Payment;
 import kr.hhplus.be.server.payment.domain.PaymentStatus;
-import kr.hhplus.be.server.payment.infrastructure.PaymentJpaRepository;
+import kr.hhplus.be.server.payment.infrastructure.repository.PaymentJpaRepository;
 import kr.hhplus.be.server.testdata.PaymentTestDataLoader;
-import kr.hhplus.be.server.user.domain.User;
 import kr.hhplus.be.server.user.infrastructure.UserJpaRepository;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
