+++ src/test/java/kr/hhplus/be/server/testdata/CouponTestDataLoader.java
@@ -1,7 +1,7 @@
 package kr.hhplus.be.server.testdata;
 
 import kr.hhplus.be.server.coupon.domain.CouponPolicy;
-import kr.hhplus.be.server.coupon.infrastructure.CouponPolicyJpaRepository;
+import kr.hhplus.be.server.coupon.infrastructure.repository.CouponPolicyJpaRepository;
 import kr.hhplus.be.server.user.domain.User;
 import kr.hhplus.be.server.user.infrastructure.UserJpaRepository;
 import lombok.RequiredArgsConstructor;
