+++ src/test/java/kr/hhplus/be/server/testdata/OrderTestDataLoader.java
@@ -3,8 +3,8 @@ package kr.hhplus.be.server.testdata;
 import kr.hhplus.be.server.order.domain.Order;
 import kr.hhplus.be.server.order.domain.OrderItem;
 import kr.hhplus.be.server.order.domain.OrderStatus;
-import kr.hhplus.be.server.order.infrastructure.OrderItemJpaRepository;
-import kr.hhplus.be.server.order.infrastructure.OrderJpaRepository;
+import kr.hhplus.be.server.order.infrastructure.repository.OrderItemJpaRepository;
+import kr.hhplus.be.server.order.infrastructure.repository.OrderJpaRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.ApplicationArguments;
 import org.springframework.boot.ApplicationRunner;
